Revert "broken routes"
