--- client/src/utils/GlobalState.js
@@ -1,17 +1,0 @@
-import React, { createContext, useContext } from "react";
-// import { useDishReducer } from './reducers'
-
-const StoreContext = createContext();
-const { Provider } = StoreContext;
-
-const StoreProvider = ({ value = [], ...props }) => {
-
-
-//   return <Provider value={[state, dispatch]} {...props} />;
-};
-
-const useStoreContext = () => {
-  return useContext(StoreContext);
-};
-
-export { StoreProvider, useStoreContext };
