+++ package.json
@@ -20,8 +20,7 @@
     "heatmap.js": "^2.0.5",
     "heatmap.js-fixed": "^2.0.2",
     "mongoose": "^6.0.13",
-    "react-bootstrap": "^2.4.0",
-    "react-router-dom": "^6.3.0"
+    "react-bootstrap": "^2.4.0"
   },
   "devDependencies": {
     "nodemon": "^2.0.3"
